+++ src/components/SocialNetworks.js
@@ -3,7 +3,8 @@ import { Grid } from '@mui/material';
 import TwitterIcon from '@mui/icons-material/Twitter';
 import { IconButton } from '@mui/material';
 import GitHubIcon from '@mui/icons-material/GitHub';
-import { FaDeviantart, FaGithub, FaTwitter } from 'react-icons/fa';
+import { FaDeviantart, FaGithub, FaTwitter, FaMastodon, Fa500Px, FaInstagram } from 'react-icons/fa';
+
 
 export default function SocialNetworks() {
 
@@ -12,13 +13,20 @@ export default function SocialNetworks() {
             <IconButton lg={12} target="_blank" href="https://twitter.com/teko_fx">
                 <FaTwitter />
             </IconButton>
-
             <IconButton target="_blank" href="https://github.com/tekofx">
                 <FaGithub />
             </IconButton>
-
             <IconButton target="_blank" href="https://www.deviantart.com/tekofx">
                 <FaDeviantart />
+            </IconButton>
+            <IconButton target="_blank" href="https://meow.social/web/@teko">
+                <FaMastodon />
+            </IconButton>
+            <IconButton target="_blank" href="https://500px.com/p/tekofx">
+                <Fa500Px />
+            </IconButton>
+            <IconButton target="_blank" href="https://instagram.com/teko_fx">
+                <FaInstagram />
             </IconButton>
 
         </>